Logica de cambio de sprite segun energia y oxigeno del personaje, resta activar cambios de salud y cantidad de oxigeno

+++ TGC.Group/Model/Personaje.cs
@@ -68,7 +68,7 @@ namespace TGC.Group.Model
             }
         }
 
-        public float Health { get => health; }
-        public float Oxygen { get => oxygen; }
+        public int Health { get => health; }
+        public int Oxygen { get => oxygen; }
     }
 }
